requirements.txt
tabula-py # This needs a Java install and JAVA_HOME environment variable set correctly
PyMuPDF